+++ CHANGELOG.md
@@ -964,5 +964,7 @@ This major version introduces the concept of `Engines` which handle the chunking
 - Output:
   - New output class: `MultipleFarmsOutput`, splitting and aggregating farm results for sub-farms and clusters
   - The `FarmLayoutOutput` can now also color by `wind_farm` and `cluster` fields.
+- Tests:
+  - New unit tests for turbine models
 
 **Full Changelog**: [https://github.com/FraunhoferIWES/foxes/commits/v1.7.7](https://github.com/FraunhoferIWES/foxes/commits/v1.7.7)